Another step towards refactor - test_forecast.py passes
